% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/typed.R
\name{typed}
\alias{typed}
\title{type}
\usage{
typed(type, typeSpeed = 0, contentType = "html", offset = "10\%",
  width = NULL, height = NULL, elementId = NULL)
}
\arguments{
\item{type}{Sentences to type, must be character vector.}

\item{typeSpeed}{Typing speed defaults to \code{0}.}

\item{contentType}{Content type, defaults to \code{html} can also be set to \code{text}.}

\item{offset}{By default the typing triggers when \code{10\%} of the element
hits the top of the window. Can also be set to pixels (\code{px}), see details and note.}

\item{width, height}{Must be a valid CSS unit (like \code{'100\%'},
\code{'400px'}, \code{'auto'}) or a number, which will be coerced to a
string and have \code{'px'} appended.}

\item{elementId}{Use specific element id.}
}
\description{
jQuery plugin that types.
}
\details{
The widget includes jQuery waypoints.js, the offset argument is passed to
to the waypoint: \url{http://imakewebthings.com/waypoints/guides/getting-started/}. See note.
}
\note{
The offset defaulting to \code{10\%} is a workaround to have it work
in RStudio viewer. It should not be a problem in Rmarkdown unless your
widget is placed at the very bottom of your document.
}
\examples{
\dontrun{
typed(c("Hello", "world"))
typed("Emphasis word <span style ='color: red;'>with html</span>.",
      contentType = "html")
typed(c("First sentence.", "Second sentence"), typeSpeed = 2)
}

}

